Use MarketplaceService:UserOwnsGamePassAsync
As documented in https://developer.roblox.com/api-reference/function/GamePassService/PlayerHasPass
PlayerHasPass will **not work with _new_ gamepasses**. Old gamepasses, created before this change, will still work for the time being however.

We can use [MarketplaceService.UserOwnsGamePassAsync(userId, gamePassId)](https://developer.roblox.com/api-reference/function/MarketplaceService/UserOwnsGamePassAsync) instead for new gamepasses.
As explained in [this public update announcement](https://devforum.roblox.com/t/live-changes-to-game-passes/116918/4), this works for old gamepasses as well, but ONLY if the gamepassID is given instead of the asset ID. This means it is not a drop in replacement for old scripts using old gamepasses via assetid. Luckily, GamePass:GamePassId is a new setting, so most people using it is are likely using the gamepassid.

I did **not** change the donners function, since I am assuming it is an old gamepass with an assetid, and that we have yet to update it to the gamepassid.